+++ crates/fuzzy_filter/src/line_splitter.rs
@@ -9,7 +9,7 @@ pub fn fuzzy_indices_fzy(line: &str, query: &str) -> ScorerOutput {
 }
 
 #[inline]
-pub(super) fn apply_skim_on_grep_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_grep_line_skim(line: &str, query: &str) -> ScorerOutput {
     strip_grep_filepath(line).and_then(|(truncated_line, offset)| {
         fuzzy_indices_skim(truncated_line, query)
             .map(|(score, indices)| (score, indices.into_iter().map(|x| x + offset).collect()))
@@ -17,7 +17,7 @@ pub(super) fn apply_skim_on_grep_line(line: &str, query: &str) -> ScorerOutput {
 }
 
 #[inline]
-pub(super) fn apply_fzy_on_grep_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_grep_line_fzy(line: &str, query: &str) -> ScorerOutput {
     strip_grep_filepath(line).and_then(|(truncated_line, offset)| {
         fuzzy_indices_fzy(truncated_line, query)
             .map(|(score, indices)| (score, indices.into_iter().map(|x| x + offset).collect()))
@@ -25,7 +25,7 @@ pub(super) fn apply_fzy_on_grep_line(line: &str, query: &str) -> ScorerOutput {
 }
 
 #[inline]
-pub(super) fn apply_substr_on_grep_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_grep_line_substr(line: &str, query: &str) -> ScorerOutput {
     strip_grep_filepath(line).and_then(|(truncated_line, offset)| {
         substr_indices(truncated_line, query)
             .map(|(score, indices)| (score, indices.into_iter().map(|x| x + offset).collect()))
@@ -33,7 +33,7 @@ pub(super) fn apply_substr_on_grep_line(line: &str, query: &str) -> ScorerOutput
 }
 
 #[inline]
-pub(super) fn apply_skim_on_file_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_file_line_skim(line: &str, query: &str) -> ScorerOutput {
     file_name_only(line).and_then(|(truncated_line, offset)| {
         fuzzy_indices_skim(truncated_line, query)
             .map(|(score, indices)| (score, indices.into_iter().map(|x| x + offset).collect()))
@@ -41,7 +41,7 @@ pub(super) fn apply_skim_on_file_line(line: &str, query: &str) -> ScorerOutput {
 }
 
 #[inline]
-pub(super) fn apply_fzy_on_file_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_file_line_fzy(line: &str, query: &str) -> ScorerOutput {
     file_name_only(line).and_then(|(truncated_line, offset)| {
         fuzzy_indices_fzy(truncated_line, query)
             .map(|(score, indices)| (score, indices.into_iter().map(|x| x + offset).collect()))
@@ -49,7 +49,7 @@ pub(super) fn apply_fzy_on_file_line(line: &str, query: &str) -> ScorerOutput {
 }
 
 #[inline]
-pub(super) fn apply_substr_on_file_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_file_line_substr(line: &str, query: &str) -> ScorerOutput {
     file_name_only(line).and_then(|(truncated_line, offset)| {
         substr_indices(truncated_line, query)
             .map(|(score, indices)| (score, indices.into_iter().map(|x| x + offset).collect()))
@@ -57,17 +57,17 @@ pub(super) fn apply_substr_on_file_line(line: &str, query: &str) -> ScorerOutput
 }
 
 #[inline]
-pub(super) fn apply_skim_on_tag_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_tag_line_skim(line: &str, query: &str) -> ScorerOutput {
     tag_name_only(line).and_then(|tag_name| fuzzy_indices_skim(tag_name, query))
 }
 
 #[inline]
-pub(super) fn apply_fzy_on_tag_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_tag_line_fzy(line: &str, query: &str) -> ScorerOutput {
     tag_name_only(line).and_then(|tag_name| fuzzy_indices_fzy(tag_name, query))
 }
 
 #[inline]
-pub(super) fn apply_substr_on_tag_line(line: &str, query: &str) -> ScorerOutput {
+pub(super) fn apply_on_tag_line_substr(line: &str, query: &str) -> ScorerOutput {
     tag_name_only(line).and_then(|tag_name| substr_indices(tag_name, query))
 }
 
@@ -80,7 +80,7 @@ mod tests {
         let query = "rules";
         let line = "crates/maple_cli/src/lib.rs:2:1:macro_rules! println_json {";
         let (_, origin_indices) = fuzzy_indices_fzy(line, query).unwrap();
-        let (_, indices) = apply_fzy_on_grep_line(line, query).unwrap();
+        let (_, indices) = apply_on_grep_line_fzy(line, query).unwrap();
         assert_eq!(origin_indices, indices);
     }
 
@@ -89,7 +89,7 @@ mod tests {
         let query = "lib";
         let line = "crates/extracted_fzy/src/lib.rs";
         let (_, origin_indices) = fuzzy_indices_fzy(line, query).unwrap();
-        let (_, indices) = apply_fzy_on_file_line(line, query).unwrap();
+        let (_, indices) = apply_on_file_line_fzy(line, query).unwrap();
         assert_eq!(origin_indices, indices);
     }
 }