+++ crates/icon/update_constants.py
@@ -31,6 +31,15 @@ with open('exactmatch_map.json', 'r') as f:
     lines.append('pub static EXACTMATCH_ICON_TABLE: &[(&str, char)] = &[%s];' %
                  joined_tuples)
 
+with open('tagkind_map.json', 'r') as f:
+    disordered = json.load(f)
+    sorted_dict = {k: disordered[k] for k in sorted(disordered)}
+
+    joined_tuples = ','.join(
+        map(lambda kv: '("%s", \'%s\')' % (kv[0], kv[1]), sorted_dict.items()))
+    lines.append('pub static TAGKIND_ICON_TABLE: &[(&str, char)] = &[%s];' %
+                 joined_tuples)
+
 lines.append('''
 pub fn bsearch_icon_table(c: &str, table: &[(&str, char)]) ->Option<usize> {
     table.binary_search_by(|&(key, _)| key.cmp(&c)).ok()