+++ crates/maple_cli/src/app.rs
@@ -36,8 +36,8 @@ pub enum Cmd {
     #[structopt(name = "ripgrep-forerunner")]
     RipGrepForerunner(crate::cmd::grep::RipGrepForerunner),
     /// Retrive the latest remote release info.
-    #[structopt(name = "check-release")]
-    CheckRelease(crate::cmd::check_release::CheckRelease),
+    #[structopt(name = "upgrade")]
+    Upgrade(upgrade::Upgrade),
 }
 
 #[derive(StructOpt, Debug)]
@@ -82,12 +82,11 @@ impl Maple {
             crate::logger::init(log_path)?;
         }
         match self.command {
-            Cmd::Version => unreachable!(),
-            Cmd::CheckRelease(_) => unreachable!(),
+            Cmd::Version | Cmd::Upgrade(_) => unreachable!(),
             Cmd::Helptags(helptags) => helptags.run()?,
-            Cmd::Tags(tags) => tags.run(self.no_cache)?,
+            Cmd::Tags(tags) => tags.run(self.no_cache, self.icon_painter)?,
             Cmd::RPC => {
-                crate::cmd::rpc::run_forever(std::io::BufReader::new(std::io::stdin()));
+                stdio_server::run_forever(std::io::BufReader::new(std::io::stdin()));
             }
             Cmd::Blines(blines) => {
                 blines.run(self.number, self.winwidth)?;