+++ crates/maple_cli/src/cmd/tags.rs
@@ -1,5 +1,5 @@
 use crate::cmd::cache::{cache_exists, send_response_from_cache, CacheEntry, SendResponse};
-use crate::ContentFiltering;
+use crate::LineSplitter;
 use anyhow::Result;
 use fuzzy_filter::{subprocess, Source};
 use itertools::Itertools;
@@ -97,14 +97,20 @@ fn create_tags_cache(args: &[&str], dir: &PathBuf) -> Result<(PathBuf, usize)> {
 }
 
 impl Tags {
-    pub fn run(&self, no_cache: bool) -> Result<()> {
+    pub fn run(&self, no_cache: bool, icon_painter: Option<icon::IconPainter>) -> Result<()> {
+        // In case of passing an invalid icon-painter option.
+        let icon_painter = icon_painter.map(|_| icon::IconPainter::ProjTags);
+
         let mut cmd_args = BASE_TAGS_ARGS.to_vec();
+
         let lang = if let Some(ref languages) = self.languages {
             format!("--languages={}", languages)
         } else {
             String::from("")
         };
+
         cmd_args.push(&lang);
+
         if self.forerunner {
             let (cache, total) = if no_cache {
                 create_tags_cache(&cmd_args, &self.dir)?
@@ -113,7 +119,7 @@ impl Tags {
             } else {
                 create_tags_cache(&cmd_args, &self.dir)?
             };
-            send_response_from_cache(&cache, total, SendResponse::Json, None);
+            send_response_from_cache(&cache, total, SendResponse::Json, icon_painter);
             return Ok(());
         } else {
             crate::cmd::filter::dynamic::dyn_fuzzy_filter_and_rank(
@@ -122,8 +128,8 @@ impl Tags {
                 None,
                 Some(30),
                 None,
-                None,
-                ContentFiltering::TagNameOnly,
+                icon_painter,
+                LineSplitter::TagNameOnly,
             )?;
         }